bears/general: Add IndentationBear
It is a generic indent bear, which looks for a start and end
indent, example: ``{ : }`` where "{" is the start indent and
"}" is the end indent.

It does not however support hanging indents or absolute indents
of any sort, also indents based on keywords are not supported yet.

for example:

	if(something)
	does not get indented

undergoes no change.